move /internal/api/cloudigrade/v1/sources_kafka/ to /internal/sources_kafka/
